refactor: create Input and PasswordInput component and add them to signin screen
